chore(exex): move notification to types crate
