+++ crates/exex/types/Cargo.toml
@@ -12,4 +12,6 @@ description = "Commonly used types for exex usage in reth."
 workspace = true
 
 [dependencies]
+reth-provider.workspace = true
+
 alloy-primitives.workspace = true
